requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
fastapi-socketio
python-socketio[asyncio]
speechbrain
torch
numpy
soundfile
scipy
python-multipart
google-cloud-speech>=2.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
librosa
psycopg2-binary